Terminado las sesiones y comprobado que funcionan

+++ layout/header.php
@@ -25,26 +25,40 @@
 
 
         <!-- Elementos del menú colapsable -->
-
+        
         <div class="collapse navbar-collapse" id="menu">
             <ul class="navbar-nav me-auto">
                 <li class="nav-item">
                     <a class="nav-link active" href="">Inventario</a>
                 </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="php/agregar.php">Agregar</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="php/eliminar.php">Eliminar</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="php/editar.php">Modificar</a>
-                </li>
+                <?php
+                    if($_SESSION['permisos'] == "admin") {
+                        echo "
+                        <li class='nav-item'>
+                            <a class='nav-link' href='php/agregar.php'>Agregar</a>
+                        </li>
+                        <li class='nav-item'>
+                            <a class='nav-link' href='php/eliminar.php'>Eliminar</a>
+                        </li>
+                        <li class='nav-item'>
+                            <a class='nav-link' href='php/editar.php'>Modificar</a>
+                        </li>";
+                    }
+                ?>
             </ul>
 
-            <!-- Boton de inicio de sesion -->
+            <!-- Boton de inicio de sesion/cerrar sesion -->
+
             <div class="d-flex">
-                <a class="d-md-inline-block btn btn-outline-warning m-auto" tabindex="-1" role="button" href="php/iniciar_sesion.php">Iniciar Sesión</a>
+                <?php
+                    if($_SESSION['usuario'] != "invitado") {
+                        echo "<p class='text-white my-auto me-2'>Usuario, {$_SESSION['usuario']}</p>
+                            <a class='d-md-inline-block btn btn-outline-warning m-auto' tabindex='-1' role='button' href='php/cerrar_sesion.php'>Cerrar sesión</a>";
+                    } else {
+                        echo "<p class='text-white my-auto me-2'>Usuario, {$_SESSION['usuario']}</p>
+                            <a class='d-md-inline-block btn btn-outline-warning m-auto' tabindex='-1' role='button' href='php/iniciar_sesion.php'>Iniciar Sesión</a>";
+                    }
+                ?>
             </div>
            
         </div>
